Complete;y working SnabbSwitch integration! We use C function for traffic processing inside Snabb's library
