--- capturecallback.c
@@ -1,92 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <stdint.h>
-#include <pthread.h>
-#include <string.h>
-#include <signal.h>
-#include <unistd.h>
-
-/* Called once before processing packets. */
-void firehose_start(); /* optional */
-
-/* Called once after processing packets. */
-void firehose_stop();  /* optional */
-
-void firehose_stop() {
-
-}
-
-/*
- * Process a packet received from a NIC.
- *
- * pciaddr: name of PCI device packet is received from
- * data:    packet payload (ethernet frame)
- * length:  payload length in bytes
- */
-inline void firehose_packet(const char *pciaddr, char *data, int length);
-
-/* Intel 82599 "Legacy" receive descriptor format.
- * See Intel 82599 data sheet section 7.1.5.
- * http://www.intel.com/content/dam/www/public/us/en/documents/datasheets/82599-10-gbe-controller-datasheet.pdf
- */
-struct firehose_rdesc {
-  uint64_t address;
-  uint16_t length;
-  uint16_t cksum;
-  uint8_t status;
-  uint8_t errors;
-  uint16_t vlan;
-} __attribute__((packed));
-
-/* Traverse the hardware receive descriptor ring.
- * Process each packet that is ready.
- * Return the updated ring index.
- */
-int firehose_callback_v1(const char *pciaddr,
-                         char **packets,
-                         struct firehose_rdesc *rxring,
-                         int ring_size,
-                         int index) {
-  while (rxring[index].status & 1) {
-    int next_index = (index + 1) & (ring_size-1);
-    __builtin_prefetch(packets[next_index]);
-    firehose_packet(pciaddr, packets[index], rxring[index].length);
-    rxring[index].status = 0; /* reset descriptor for reuse */
-    index = next_index;
-  }
-  return index;
-}
-
-
-uint64_t received_packets = 0;
-
-void* speed_printer(void* ptr) {
-    while (1) {
-        uint64_t packets_before = received_packets;
-    
-        sleep(1);
-    
-        uint64_t packets_after = received_packets;
-        uint64_t pps = packets_after - packets_before;
- 
-        printf("We process: %llu pps\n", pps);
-    }   
-}
-
-void sigproc(int sig) {
-    firehose_stop();
-
-    printf("We caught SINGINT and will finish application\n");
-    exit(0);
-}
-
-
-void firehose_start() {
-    pthread_t thread;
-    pthread_create(&thread, NULL, speed_printer, NULL);
-    pthread_detach(thread);
-}
-
-void firehose_packet(const char *pciaddr, char *data, int length) {
-    __sync_fetch_and_add(&received_packets, 1);
-}
